Crate test_cluster

Structs

FullNodeHandle

PcoolFlowOverride

Reverts the P-COOL protocol flag override when dropped.

RandomNodeRestarter

TestCluster

TestClusterBuilder

Functions

override_pcool_flow

Sets the P-COOL protocol flag for every node of a test cluster built while the returned guard is alive.